stock-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
httpx
sqlalchemy
databases[asyncpg]
asyncpg
alembic
aiofiles
Jinja2
email-validator
passlib[bcrypt]
python-multipart
pytest
pytest-asyncio
firebase-admin
yfinance